fixing some translations and removing some traces

+++ app/assets/javascripts/nabu_editor/dashboard.js
@@ -19,7 +19,7 @@ filter_table
 mount_point
 */
 
-// neccesary variables 
+// neccesary variables
 // TODO, make these NOT neccessary
 
 var agent;
@@ -40,14 +40,14 @@ var smallvideo = false;
 var visible_container;
 var wasReallyPlaying = false
 
-// REACT reference  
+// REACT reference
 window.filter_table = "not set yet";
 
 // ### Overrides
 // override function nabu/assets/javasctipt/marduq/player.js
 var popOnLoadStart = function() {
   console.log("POP ON LOAD START")
-  $('#video_frame').fadeIn('slow'); 
+  $('#video_frame').fadeIn('slow');
   // failsafe
   setTimeout( function() { $('#video_frame').css('opacity', 1)}, 800 )
   if (!describe_initialized ) initDescribe();
@@ -57,9 +57,9 @@ var popOnLoadStart = function() {
 // override function nabu/assets/javasctipt/marduq/player.js
 var popCanPlay = function () {
   console.log("POP CAN PLAY")
-  $('#nabu_controls').fadeIn('slow');   
-  initControls(); 
-  controlsAreVisible = true;  
+  $('#nabu_controls').fadeIn('slow');
+  initControls();
+  controlsAreVisible = true;
   controls_initialized = true;
   $('.playhead').css('width','0');
   margin_select_video();
@@ -69,16 +69,16 @@ var doSelectMovie = function(id) {
   setProgram( id );
   if ( $('.video_uploader_container').is(':visible') ) showCreateMovie();
   $(".contentwrapper").animate({ scrollTop: 0 }, "slow");
-  setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );  
+  setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );
 };
 
-var doShowMovie = function(id) {  
+var doShowMovie = function(id) {
   setProgram( id );
   if ( $('.video_uploader_container').is(':visible') ) showCreateMovie();
-  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();  
-  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();  
+  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();
+  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();
   if ( $('.editor_container').is(':visible') ) showSettingsMovie();
-  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();  
+  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();
   $(".contentwrapper").animate({ scrollTop: 0 }, "slow");
   setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );
 };
@@ -86,10 +86,10 @@ var doShowMovie = function(id) {
 var doDescribeMovie = function(id) {
   setProgram( id );
   if ( $('.video_uploader_container').is(':visible') ) showCreateMovie();
-  if ( !$('.video_describe_container').is(':visible') ) showDescribeMovie();  
-  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();  
+  if ( !$('.video_describe_container').is(':visible') ) showDescribeMovie();
+  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();
   if ( $('.editor_container').is(':visible') ) showSettingsMovie();
-  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();  
+  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();
   $(".contentwrapper").animate({ scrollTop: 0 }, "slow");
   //showDescribeMovie();
   setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );
@@ -98,10 +98,10 @@ var doDescribeMovie = function(id) {
 var doTimeLineEditMovie = function(id) {
   setProgram( id );
   if ( $('.video_uploader_container').is(':visible') ) showCreateMovie();
-  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();  
-  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();  
+  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();
+  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();
   if ( !$('.editor_container').is(':visible') ) showSettingsMovie();
-  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();  
+  if ( $('.video_publish_container').is(':visible') ) showPublishMovie();
   $(".contentwrapper").animate({ scrollTop: 0 }, "slow");
   //showSettingsMovie();
   setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );
@@ -110,19 +110,19 @@ var doTimeLineEditMovie = function(id) {
 var doPublishMovie = function(id) {
   setProgram( id );
   if ( $('.video_uploader_container').is(':visible') ) showCreateMovie();
-  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();  
-  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();  
+  if ( $('.video_describe_container').is(':visible') ) showDescribeMovie();
+  //if ( $('.describe_movie_advanced').is(':visible') ) showDescribeMovie();
   if ( $('.editor_container').is(':visible') ) showSettingsMovie();
-  if ( !$('.video_publish_container').is(':visible') ) showPublishMovie();  
+  if ( !$('.video_publish_container').is(':visible') ) showPublishMovie();
   $(".contentwrapper").animate({ scrollTop: 0 }, "slow");
-  //showPublishMovie();  
+  //showPublishMovie();
   setTimeout( function() { if ( $('.leprograms').is(":visible") ) $('.leprograms').shapeshift( shapeshiftOptions() ) }, 100 );
 };
 
 var doCopyMovie = function(id) {
   $.ajax({
     url: '/marduq_api/programs/copy/',
-    type: 'POST',    
+    type: 'POST',
     data: {'program': id },
     success: function() { setPrograms() },
     done: function() {}
@@ -130,13 +130,15 @@ var doCopyMovie = function(id) {
 }
 
 var doDeleteMovie = function(id) {
-  setProgram(-1);
-  $.ajax({
-    url: '/marduq_api/programs/' + id,
-    type: 'DELETE',
-    success: function() { setPrograms() },
-    done: function() {}
-  });
+  if ( confirm( t.user_interface.confirm_delete_program )  ) {
+    setProgram(-1);
+    $.ajax({
+      url: '/marduq_api/programs/' + id,
+      type: 'DELETE',
+      success: function() { setPrograms() },
+      done: function() {}
+    });
+  }
 };
 
 //Slide down videos if NABU_CONTROLS
@@ -158,7 +160,7 @@ var resize_aspect_ration_change = function() {
   aspect_ratio = $('.dropdown_select_ratio').val();
   var arr = aspect_ratio.split(':');
   aspect_width = arr[0];
-  aspect_height = arr[1];  
+  aspect_height = arr[1];
   resize_aspect_ratio();
 }
 
@@ -166,42 +168,42 @@ setTimeout( margin_select_video(), 800 );
 
 // ### Helpers
 // update current program
-var setProgram = function( id ) {  
+var setProgram = function( id ) {
   console.log("set program: ", id )
-  if ( program !== null && program.id === id ) return;      
-  
+  if ( program !== null && program.id === id ) return;
+
   $('#video_frame').fadeOut('fast');
   $('#nabu_controls').fadeOut('slow');
   $('.big-play').addClass('hidden');
   $('#video_frame').animate({'opacity':'1'});
   setTimeout(margin_select_video(), 500);
-  
+
   if ( id == -1 || id == "-1") {
     $('.video_uploader_container').fadeOut('slow');
     $('.playhead').css('width','0');
     return;
-  }     
-  
+  }
+
   // add id to url
   window.location.hash = "i" + id;
-  
+
   // retreive the program from the server
-  
+
   // #######################################################
   // ### THIS NEEDS TO BE FUCKING SCOPED with CLIENT ID !!!!
   // #######################################################
-  
+
   $.get('/marduq_api/programs/'+id, function(p) {
-    program = p;    
-    getPlayer( program, "#video_frame", agent.technology, agent.videotype, null );            
+    program = p;
+    getPlayer( program, "#video_frame", agent.technology, agent.videotype, null );
     controlsAreVisible = false;
     setMarqerAfterPopInits();
     setDataFromProgram();
     setDisabledStatesLeftMenu();
     setSocial()
     if(!smallvideo) {
-      setTimeout( resize_aspect_ratio, 200);   
-    } 
+      setTimeout( resize_aspect_ratio, 200);
+    }
   });
 };
 
@@ -227,21 +229,21 @@ function setDisabledStatePublish() {
 }
 
 // wait for pop and set marqers
-function setMarqerAfterPopInits() {  
+function setMarqerAfterPopInits() {
   if ( pop === null || isNaN( pop.duration() ) ) {
     setTimeout( function( mrqs ) { setMarqerAfterPopInits(); setDisabledStatesLeftMenu(); }, 300 );
   }else{
     resetMarqers();
     setMarqers( program.marqers, pop.duration() );
-    convertMarqersIntoTrackevents()    
+    convertMarqersIntoTrackevents()
   }
 }
 
 // update all programs
 function setPrograms() {
   console.log( "Update Programs: /editor3/get_programs/" );
-  
-  $.get( '/'+mount_point+'/get_programs/', function( resp ) {    
+
+  $.get( '/'+mount_point+'/get_programs/', function( resp ) {
     // update react
     console.log( "Update programs Succes" );
     console.log( resp );
@@ -249,67 +251,62 @@ function setPrograms() {
     updatePrograms( resp ) // for edit_menu.js, or the publish menu list
 
     // reset the selected program, if it was updating
-    setTimeout( function() { 
-      $('.leprograms').find('.selected').animate({'opacity':1}, 200); 
+    setTimeout( function() {
+      $('.leprograms').find('.selected').animate({'opacity':1}, 200);
       setDisabledStatePublish();
     }, 500 );
   });
-  
+
 }
 
 // update shapeshift
-var shapeshiftOptions = function() {  
+var shapeshiftOptions = function() {
   var a_w = $('.video_search_wrapper').width();
   var c_w = $('.program_container').width();
   var numColums = Math.floor( a_w / c_w );
   var roomLeft = a_w - ( c_w * numColums );
-  var gutterX = Math.floor( roomLeft / ( numColums - 1 ) ); 
+  var gutterX = Math.floor( roomLeft / ( numColums - 1 ) );
   return { enableDrag: false, gutterX: gutterX, gutterY: (gutterX/2), paddingX: 20 };
 };
 
 // keep video aspect ratio
 function resize_aspect_ratio() {
   console.log("resize aspect ", $('#video_frame').hasClass('fullscreen'))
-  
+
   // we're in full screen
   if ( $('#video_frame').hasClass('fullscreen') ) {
     $(".videowrapper").css({"height": "100%"});
     $(".videowrapper").css({"width": "100%"});
     $(".videowrapper").css({"margin": "0"});
-    
-    $(".video_holder").css({"height": "100%"});    
+
+    $(".video_holder").css({"height": "100%"});
     $(".ui-resizable-se").hide()
     return;
   }
 
-  // resets  
+  // resets
   $(".video_holder").css({"height": "initial"});
-  $(".videowrapper").css({"margin": "2% auto"});  
+  $(".videowrapper").css({"margin": "2% auto"});
   $(".ui-resizable-se").fadeIn()
 
-  console.log("allo?")
-
-  setTimeout( function() { 
-    console.log("ploing")
-    
+  setTimeout( function() {
     var width = $(".videowrapper").width();
     var height = width * ( aspect_height / aspect_width );
     var fixedheight = height //.toFixed(0);
-    
-    $(".videowrapper").animate({"height": (fixedheight)+ 'px'}, 300);    
-    console.log('wat nou:', $('.contentwrapper').width(),  width )
-    
-    if ( $('.contentwrapper').width() < width ) $(".videowrapper").css({"width": ""}); // reset        
+
+    $(".videowrapper").animate({"height": (fixedheight)+ 'px'}, 300);
+
+    if ( $('.contentwrapper').width() < width ) $(".videowrapper").css({"width": ""}); // reset
     //updateMarqerBottomControls(fixedheight);
 
   }, 600 );
-  
+
   // failsafe for laggy servers (looking at you kaltura)
-  setTimeout( function() { 
-    if ( $('#video_frame').css('display') == 'none' ) {      
+  setTimeout( function() {
+    if ( $('#video_frame').css('display') == 'none' ) {
       $('#video_frame').fadeIn('slow');
     }
-  }, 500 );  
+  }, 500 );
 }
 
 //function updateMarqerBottomControls( fh ) {
@@ -324,7 +321,7 @@ function checkURLForPresets() {
     setProgram( window.location.hash.substring(2) );
     showCreateMovie( false );
   }
-  
+
   // if #m open marqer editor ?
   //if ( window.location.hash.substring(0,2) == '#m' ) {
 }
@@ -334,24 +331,24 @@ var original_width = $(document).width();
 $(document).ready(function() {
   //set var aspect_ratio
   console.log('ready aspect ratio' + aspect_ratio);
-  
+
   $('.logopreview img').change(function(){
     setLogoHeight();
   });
-  
+
   setDisabledStatesLeftMenu();
   setDisabledStatePublish();
-  
+
   // check out the user
   agent = checkUserAgent();
   setTimeout( function() { margin_select_video(); }, 1000);             // fail safe
   setTimeout( function() { $('#video_frame').css('opacity', 1)}, 1500 ) // fail safer
-  
+
   // Keep aspect ratio
   $(window).resize(function() {
 
     // always move editor
-    //updateMarqerBottomControls() 
+    //updateMarqerBottomControls()
 
     // Quickfix
     if ( original_width == $(document).width() ) return;
@@ -363,31 +360,31 @@ $(document).ready(function() {
     var right_menu_width = $(".right_menu").width();
     var content_wrapper_width = document_width - nav_width - right_menu_width;
     $(".contentwrapper").css("width", content_wrapper_width.toFixed(0) - 2);
-  
+
     // Keep shapeshifter
     setTimeout( function() { if ( $('.leprograms').is(":visible") ) try { $('.leprograms').shapeshift(shapeshiftOptions())} catch(e){}} , 800 );
   }).resize();
-  
+
   // Initialize Programs
   setTimeout( function() { try { $('.leprograms').shapeshift(shapeshiftOptions()); $('.leprograms').fadeIn('slow'); } catch(e){} } , 800 );
-  
+
   // Initialize big-play
   $('#nabu_controls').hide();
-  $('.big-play').click( function() { 
-    $('.big-play').addClass('hidden');    
+  $('.big-play').click( function() {
+    $('.big-play').addClass('hidden');
     doControl('play');
     setTimeout( function() { margin_select_video() }, 800);
   });
-  
+
   // TODO: also on swipe up and down
   // rewrite this with a class (?)
   $('#butt-up').click( animateVideoUp );
   $('#butt-down').click( animateVideoDown );
   //$('.video_holder').on('swipedown',function(){alert("swipedown..");} );
   //$('.video_holder').on('swipeup',function(){alert("swipedown..");} );
-  
+
   //Material design dropdowns hides the selectlist and shows a UL with listitems
-  //The selectlist wont change 
+  //The selectlist wont change
   //This will make it to work for all of them.
   $('.dropdownjs ul li').click(function(){
       var selectedvalue = $(this).attr('value');
@@ -401,45 +398,45 @@ $(document).ready(function() {
 
   $(document).on('change','.dropdown_select_ratio',function(){
     if(!smallvideo) {
-      resize_aspect_ration_change() 
+      resize_aspect_ration_change()
     }
   });
 
-  // Hide the sections  
+  // Hide the sections
   $('.video_describe_container').hide();
   $('.video_publish_container').hide();
   $('.describe_movie_advanced').hide();
   $('.zoomContainer').hide();
   $('.channels_container').hide();
-  
+
   // Hide the helpers
   $('.describe_helper').hide();
   $('.publish_helper').hide();
-    
+
   /////////////////////////////////////////////////
   //HIDE TESTERS FOR MARQERS FOR TEST PURPOSE ONLY!
   /////////////////////////////////////////////////
   $('.marqers_container').hide();
-  
+
   //Hide the helpers when clicked
   $('.readed').click(function() {
     $(this).parent().fadeOut('fast',function(){
       $(this).addClass('hidden');
     });
   });
-  
+
   // listen for fullscreen, and discard the max height from aspect ratio
-  document.addEventListener("fullscreenchange", delayedScreenChange, false);      
+  document.addEventListener("fullscreenchange", delayedScreenChange, false);
   document.addEventListener("webkitfullscreenchange", delayedScreenChange, false);
   document.addEventListener("mozfullscreenchange", delayedScreenChange, false);
   resize_aspect_ratio()
-  
+
   // set the bottom part of the editor
   //updateMarqerBottomControls
-  
+
   // check for hash &c
   checkURLForPresets()
-  
+
   // set resizable
   $('.videowrapper').resizable({
     aspectRatio: true,
